backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Agentic Document Generator Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-xai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # PDF generation
    "xhtml2pdf>=0.2.16",
    "markdown>=3.6",
    # File upload / reference extraction
    "python-multipart>=0.0.9",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
